refactor, mving to some skycoin specific implementation ref #339
