+++ tiny-firmware/firmware/fsm.c
@@ -318,30 +318,6 @@ void fsm_msgGetFeatures(GetFeatures *msg) {
     msg_write(MessageType_MessageType_Features, resp);
 }
 
-void fsm_msgSkycoinCheckMessageSignature(SkycoinCheckMessageSignature *msg) {
-    GET_MSG_POINTER(Success, successResp);
-    GET_MSG_POINTER(Failure, failureResp);
-    uint16_t msg_id = MessageType_MessageType_Failure;
-    void *msg_ptr = failureResp;
-    switch (msgSkycoinCheckMessageSignatureImpl(msg, successResp, failureResp)) {
-        case ErrOk:
-            msg_id = MessageType_MessageType_Success;
-            msg_ptr = successResp;
-            layoutRawMessage("Verification success");
-            break;
-        case ErrAddressGeneration:
-        case ErrInvalidSignature:
-            failureResp->code = FailureType_Failure_InvalidSignature;
-            layoutRawMessage("Wrong signature");
-            break;
-        default:
-            strncpy(failureResp->message, _("Firmware error."), sizeof(failureResp->message));
-            layoutHome();
-            break;
-    }
-    msg_write(msg_id, msg_ptr);
-}
-
 ErrCode_t requestConfirmTransaction(char *strCoin, char *strHour, TransactionSign *msg, uint32_t i) {
     layoutDialogSwipe(&bmp_icon_question, _("Cancel"), _("Next"), NULL, _("Do you really want to"), strCoin, strHour,
                       _("to address"), _("..."), NULL);
@@ -349,121 +325,6 @@ ErrCode_t requestConfirmTransaction(char *strCoin, char *strHour, TransactionSig
     layoutAddress(msg->transactionOut[i].address);
     CHECK_BUTTON_PROTECT_RET_ERR_CODE
     return ErrOk;
-}
-
-void fsm_msgTransactionSign(TransactionSign *msg) {
-    CHECK_PIN
-    CHECK_MNEMONIC
-    CHECK_INPUTS(msg)
-    CHECK_OUTPUTS(msg)
-
-    MessageType msgtype = MessageType_MessageType_TransactionSign;
-    RESP_INIT(ResponseTransactionSign);
-    ErrCode_t err = msgTransactionSignImpl(msg, &requestConfirmTransaction, resp);
-    char *failMsg = NULL;
-    switch (err) {
-    case ErrOk:
-        msg_write(MessageType_MessageType_ResponseTransactionSign, resp);
-        break;
-    case ErrAddressGeneration:
-        failMsg = _("Wrong return address");
-        // fall through
-    default:
-        fsm_sendResponseFromErrCode(err, NULL, failMsg, &msgtype);
-        break;
-    }
-    layoutHome();
-}
-
-void fsm_msgSkycoinSignMessage(SkycoinSignMessage* msg)
-{
-    CHECK_MNEMONIC
-    RESP_INIT(ResponseSkycoinSignMessage);
-    CHECK_PIN_UNCACHED
-
-    MessageType msgtype = MessageType_MessageType_SkycoinSignMessage;
-    ResponseSkycoinAddress respAddr;
-    uint8_t seckey[32] = {0};
-    uint8_t pubkey[33] = {0};
-    if (msg->has_bip44_addr) {
-        const char* mnemo = storage_getFullSeed();
-        uint8_t seed[512 / 8] = {0};
-        mnemonic_to_seed(mnemo, "", seed, NULL);
-        size_t addr_size = sizeof(respAddr.addresses[0]);
-        int ret = hdnode_address_for_branch(
-            seed, sizeof(seed), msg->bip44_addr.purpose,
-            msg->bip44_addr.coin_type, msg->bip44_addr.account,
-            msg->bip44_addr.change, msg->bip44_addr.address_start_index, respAddr.addresses[0], &addr_size);
-        if (ret != 1) {
-            fsm_sendResponseFromErrCode(
-                ErrFailed, NULL, _("Unable to get address"),&msgtype);
-            layoutHome();
-            return;
-        }
-    } else {
-        ErrCode_t err = fsm_getKeyPairAtIndex(1, pubkey, seckey, &respAddr,
-                                              msg->address_n);
-        if (err != ErrOk) {
-            fsm_sendResponseFromErrCode(
-                err, NULL, _("Unable to get keys pair"),&msgtype);
-            layoutHome();
-            return;
-        }
-    }
-    layoutDialogSwipe(&bmp_icon_question, _("Cancel"), _("Confirm"), NULL,
-                      _("Do you really want to"), _("sign message using"),
-                      _("this address?"), respAddr.addresses[0], NULL, NULL);
-    CHECK_BUTTON_PROTECT
-
-    ErrCode_t err = msgSkycoinSignMessageImpl(msg, resp);
-    if (err == ErrOk) {
-        msg_write(MessageType_MessageType_ResponseSkycoinSignMessage, resp);
-        layoutRawMessage("Signature success");
-    } else {
-        char *failMsg = NULL;
-        if (err == ErrMnemonicRequired) {
-            failMsg = _("Mnemonic not set");
-        }
-        fsm_sendResponseFromErrCode(err, NULL, failMsg, &msgtype);
-        layoutHome();
-    }
-}
-
-void fsm_msgSkycoinAddress(SkycoinAddress* msg)
-{
-    MessageType msgtype = MessageType_MessageType_SkycoinAddress;
-    RESP_INIT(ResponseSkycoinAddress);
-    char *failMsg = NULL;
-    ErrCode_t err = msgSkycoinAddressImpl(msg, resp);
-    switch (err) {
-        case ErrUserConfirmation:
-            layoutAddress(resp->addresses[0]);
-            if (!protectButton(ButtonRequestType_ButtonRequest_ProtectCall, false)) {
-                err = ErrActionCancelled;
-                break;
-            }
-            // fall through
-        case ErrOk:
-            msg_write(MessageType_MessageType_ResponseSkycoinAddress, resp);
-            layoutHome();
-            return;
-        case ErrPinRequired:
-            failMsg = _("Expected pin");
-            break;
-        case ErrTooManyAddresses:
-            failMsg = _("Asking for too much addresses");
-            break;
-        case ErrMnemonicRequired:
-            failMsg = _("Mnemonic required");
-            break;
-        case ErrAddressGeneration:
-            failMsg = _("Key pair generation failed");
-            break;
-        default:
-            break;
-    }
-    fsm_sendResponseFromErrCode(err, NULL, failMsg, &msgtype);
-    layoutHome();
 }
 
 void fsm_msgPing(Ping* msg)
