+++ tiny-firmware/firmware/fsm.h
@@ -32,9 +32,6 @@ void fsm_sendFailure(FailureType code, const char* text, MessageType* msgtype);
 void fsm_msgInitialize(Initialize* msg);
 void fsm_msgGetFeatures(GetFeatures* msg);
 void fsm_msgApplySettings(ApplySettings* msg);
-void fsm_msgSkycoinCheckMessageSignature(SkycoinCheckMessageSignature* msg);
-void fsm_msgSkycoinSignMessage(SkycoinSignMessage* msg);
-void fsm_msgSkycoinAddress(SkycoinAddress* msg);
 void fsm_msgGenerateMnemonic(GenerateMnemonic* msg);
 void fsm_msgSetMnemonic(SetMnemonic* msg);
 void fsm_msgPing(Ping* msg);
@@ -50,7 +47,6 @@ void fsm_msgPinMatrixAck(PinMatrixAck* msg);
 void fsm_msgCancel(Cancel* msg);
 void fsm_msgRecoveryDevice(RecoveryDevice* msg);
 void fsm_msgWordAck(WordAck* msg);
-void fsm_msgTransactionSign(TransactionSign* msg);
 void fsm_msgSignTx(SignTx* msg);
 void fsm_msgTxAck(TxAck* msg);
